refactor: optimize docker-compose for local dev, skip registry pulls and app builds
