+++ docker-compose.yml
@@ -16,11 +16,9 @@ services:
       timeout: 5s
       retries: 5
   v6y-migrate:
-    extends:
-      file: docker-compose.base.yml
-      service: v6y-service-base
-    command: pnpm --filter @v6y/core-logic db:migrate:deploy
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-bff:latest
+    build:
+      context: .
+      dockerfile: v6y-config/Dockerfile.migrations
     environment:
       - PSQL_DB_HOST=${PSQL_DB_HOST}
       - PSQL_DB_NAME=${PSQL_DB_NAME}
@@ -31,6 +29,9 @@ services:
       v6y-database:
         condition: service_healthy
     restart: "no"
+    volumes:
+      - ./:/app
+      - mono-node-modules:/app/node_modules
 
   v6y-bff:
     extends:
@@ -39,7 +40,6 @@ services:
     command: pnpm run start:bff
     ports:
       - "4001:4001"
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-bff:latest
     environment:
       - PSQL_DB_HOST=${PSQL_DB_HOST}
       - PSQL_DB_NAME=${PSQL_DB_NAME}
@@ -67,7 +67,6 @@ services:
     command: pnpm run start:bfb:main
     ports:
       - "4002:4002"
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-bfb-main-analyzer:latest
     environment:
       - PSQL_DB_HOST=${PSQL_DB_HOST}
       - PSQL_DB_NAME=${PSQL_DB_NAME}
@@ -97,7 +96,6 @@ services:
     command: pnpm run start:bfb:static
     ports:
       - "4003:4003"
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-bfb-static-code-auditor:latest
     environment:
       - PSQL_DB_HOST=${PSQL_DB_HOST}
       - PSQL_DB_NAME=${PSQL_DB_NAME}
@@ -125,7 +123,6 @@ services:
     command: pnpm run start:bfb:dynamic
     ports:
       - "4004:4004"
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-bfb-url-dynamic-auditor:latest
     environment:
       - PSQL_DB_HOST=${PSQL_DB_HOST}
       - PSQL_DB_NAME=${PSQL_DB_NAME}
@@ -153,7 +150,6 @@ services:
     command: pnpm run start:bfb:devops
     ports:
       - "4005:4005"
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-bfb-devops-auditor:latest
     environment:
       - PSQL_DB_HOST=${PSQL_DB_HOST}
       - PSQL_DB_NAME=${PSQL_DB_NAME}
@@ -181,7 +177,6 @@ services:
     command: pnpm run start:frontend
     ports:
       - "4006:4006"
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-frontend:latest
     environment:
       - NEXT_PUBLIC_V6Y_BFF_PATH=${NEXT_PUBLIC_V6Y_BFF_PATH}
     healthcheck:
@@ -197,7 +192,6 @@ services:
     command: pnpm run start:frontend:bo
     ports:
       - "4008:4008"
-    image: ghcr.io/${REGISTRY_OWNER}/v6y-frontend-bo:latest
     environment:
       - NEXT_PUBLIC_V6Y_BFF_PATH=${NEXT_PUBLIC_V6Y_BFF_PATH}
       - NEXTAUTH_URL=${NEXTAUTH_URL}
